# docs/swagger.yaml
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /check:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 类型
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 登陆验证接口
  /message:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 来源uid
        in: formData
        name: from_id
        required: true
        type: string
      - description: 目标uid
        in: formData
        name: toID
        required: true
        type: string
      - description: 内容
        in: formData
        name: content
        required: true
        type: string
      - description: 类型|kefu,visitor
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 发送消息接口
  /messages:
    get:
      consumes:
      - multipart/form-data
      parameters:
      - description: 访客ID
        in: query
        name: visitorId
        required: true
        type: string
      - description: 认证token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取访客聊天信息接口
  /visitors:
    get:
      consumes:
      - multipart/form-data
      parameters:
      - description: 分页
        in: query
        name: page
        required: true
        type: string
      - description: 认证token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取访客列表接口
swagger: "2.0"